pyproject.toml
[tool.poetry]
name = "supabase"
version = "2.15.2" # {x-release-please-version}
description = "Supabase client for Python."
authors = ["Joel Lee <joel@joellee.org>", "Leon Fedden <leonfedden@gmail.com>", "Daniel Reinón García <danielreinon@outlook.com>", "Leynier Gutiérrez González <leynier41@gmail.com>", "Anand", "Andrew Smith <a.smith@silentworks.co.uk>"]
homepage = "https://github.com/supabase/supabase-py"
repository = "https://github.com/supabase/supabase-py"
documentation = "https://github.com/supabase/supabase-py"
readme = "README.md"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
postgrest = ">0.19,<1.1"
realtime = ">=2.4.0 <2.5.0"
gotrue = "^2.11.0"
httpx = ">=0.26,<0.29"
storage3 = ">=0.10,<0.12"
supafunc = "^0.9"
streamlit = "^1.37.0"
pandas = "^2.0.0"
plotly = "^5.17.0"
openai = "^1.0.0"
python-dotenv = "^1.1.0"

[tool.poetry.dev-dependencies]
pre-commit = "^4.1.0"
black = "^25.1"
pytest = "^8.3.5"
flake8 = "^7.2.0"
isort = "^6.0.1"
pytest-cov = "^6.1.1"
commitizen = "^4.8.2"

[tool.poetry.scripts]
tests = 'poetry_scripts:run_tests'

[tool.poetry.group.dev.dependencies]
unasync-cli = { git = "https://github.com/supabase-community/unasync-cli.git", branch = "main" }
pytest-asyncio = ">=0.24,<0.27"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
